fixed -Wincompatible-pointer-types compilation warnings onsubstdio.h

+++ qmail-pop3d.c
@@ -25,7 +25,7 @@
 
 void die() { _exit(0); }
 
-int saferead(fd,buf,len) int fd; char *buf; int len;
+ssize_t saferead(fd,buf,len) int fd; char *buf; int len;
 {
   int r;
   r = timeoutread(1200,fd,buf,len);
@@ -33,7 +33,7 @@ int saferead(fd,buf,len) int fd; char *buf; int len;
   return r;
 }
 
-int safewrite(fd,buf,len) int fd; char *buf; int len;
+ssize_t safewrite(fd,buf,len) int fd; char *buf; int len;
 {
   int r;
   r = timeoutwrite(1200,fd,buf,len);
